Lazy load item and web content for non-wifi connections
