+++ app/src/main/java/io/github/hidroh/materialistic/ReadabilityFragment.java
@@ -22,12 +22,10 @@ import javax.inject.Inject;
 
 import io.github.hidroh.materialistic.data.ReadabilityClient;
 
-public class ReadabilityFragment extends BaseFragment implements Scrollable {
+public class ReadabilityFragment extends LazyLoadFragment implements Scrollable {
     public static final String EXTRA_URL = ReadabilityFragment.class.getName() + ".EXTRA_URL";
     private static final String STATE_CONTENT = "state:content";
     private static final String STATE_TEXT_SIZE = "state:textSize";
-    private static final String STATE_EAGER_LOAD = "state:eagerLoad";
-    private static final String STATE_ACTIVITY_CREATED = "state:activityCreated";
     private NestedScrollView mScrollView;
     private TextView mTextView;
     private ProgressBar mProgressBar;
@@ -35,8 +33,6 @@ public class ReadabilityFragment extends BaseFragment implements Scrollable {
     private String mContent;
     private float mTextSize;
     private String[] mTextSizeOptionValues;
-    private boolean mEagerLoad;
-    private boolean mActivityCreated;
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
@@ -45,13 +41,10 @@ public class ReadabilityFragment extends BaseFragment implements Scrollable {
         if (savedInstanceState != null) {
             mTextSize = savedInstanceState.getFloat(STATE_TEXT_SIZE);
             mContent = savedInstanceState.getString(STATE_CONTENT);
-            mEagerLoad = savedInstanceState.getBoolean(STATE_EAGER_LOAD);
-            mActivityCreated = savedInstanceState.getBoolean(STATE_ACTIVITY_CREATED);
         } else {
             mTextSize = AppUtils.getDimension(getActivity(),
                     Preferences.Theme.resolvePreferredTextSizeResId(getActivity()),
                     R.attr.contentTextSize);
-            mEagerLoad = AppUtils.isOnWiFi(getContext());
         }
         mTextSizeOptionValues = getResources().getStringArray(R.array.pref_text_size_values);
     }
@@ -94,28 +87,10 @@ public class ReadabilityFragment extends BaseFragment implements Scrollable {
     }
 
     @Override
-    public void onActivityCreated(Bundle savedInstanceState) {
-        super.onActivityCreated(savedInstanceState);
-        mActivityCreated = true;
-        eagerLoad();
-    }
-
-    @Override
-    public void setUserVisibleHint(boolean isVisibleToUser) {
-        super.setUserVisibleHint(isVisibleToUser);
-        if (isVisibleToUser && !mEagerLoad) {
-            mEagerLoad = true;
-            eagerLoad();
-        }
-    }
-
-    @Override
     public void onSaveInstanceState(Bundle outState) {
         super.onSaveInstanceState(outState);
         outState.putFloat(STATE_TEXT_SIZE, mTextSize);
         outState.putString(STATE_CONTENT, mContent);
-        outState.putBoolean(STATE_EAGER_LOAD, mEagerLoad);
-        outState.putBoolean(STATE_ACTIVITY_CREATED, mActivityCreated);
     }
 
     @Override
@@ -123,13 +98,8 @@ public class ReadabilityFragment extends BaseFragment implements Scrollable {
         mScrollView.smoothScrollTo(0, 0);
     }
 
-    private void eagerLoad() {
-        if (!mEagerLoad) {
-            return;
-        }
-        if (!mActivityCreated) {
-            return;
-        }
+    @Override
+    protected void load() {
         if (TextUtils.isEmpty(mContent)) {
             parse();
         } else {
